requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
websockets==14.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
mongomock-motor==0.0.34
